Rename PredicateDefaults --> PredicateCommon
