+++ validation/src/main/scala/hmda/validation/rules/lar/syntactical/S010.scala
@@ -4,7 +4,7 @@ import hmda.model.fi.lar.LoanApplicationRegister
 import hmda.validation.dsl.Result
 import hmda.validation.rules.EditCheck
 import hmda.validation.dsl.PredicateSyntax._
-import hmda.validation.dsl.PredicateDefaults._
+import hmda.validation.dsl.PredicateCommon._
 
 object S010 extends EditCheck[LoanApplicationRegister] {
