+++ validation/src/main/scala/hmda/validation/rules/lar/validity/V320.scala
@@ -3,7 +3,7 @@ package hmda.validation.rules.lar.validity
 import hmda.model.fi.lar.LoanApplicationRegister
 import hmda.validation.dsl.Result
 import hmda.validation.rules.EditCheck
-import hmda.validation.dsl.PredicateDefaults._
+import hmda.validation.dsl.PredicateCommon._
 import hmda.validation.dsl.PredicateSyntax._
 
 object V320 extends EditCheck[LoanApplicationRegister] {
