+++ validation/src/main/scala/hmda/validation/rules/lar/validity/V330.scala
@@ -5,7 +5,7 @@ import hmda.validation.dsl.{ Failure, Result }
 import hmda.validation.rules.EditCheck
 
 import scala.util.Try
-import hmda.validation.dsl.PredicateDefaults._
+import hmda.validation.dsl.PredicateCommon._
 import hmda.validation.dsl.PredicateSyntax._
 
 object V330 extends EditCheck[LoanApplicationRegister] {
