+++ validation/src/main/scala/hmda/validation/rules/lar/validity/V470.scala
@@ -4,7 +4,7 @@ import hmda.model.fi.lar.LoanApplicationRegister
 import hmda.validation.dsl.Result
 import hmda.validation.rules.EditCheck
 import scala.util.Try
-import hmda.validation.dsl.PredicateDefaults._
+import hmda.validation.dsl.PredicateCommon._
 import hmda.validation.dsl.PredicateSyntax._
 
 object V470 extends EditCheck[LoanApplicationRegister] {
