+++ validation/src/main/scala/hmda/validation/rules/ts/syntactical/S013.scala
@@ -3,7 +3,7 @@ package hmda.validation.rules.ts.syntactical
 import hmda.model.fi.ts.TransmittalSheet
 import hmda.validation.dsl.Result
 import scala.concurrent.{ ExecutionContext, Future }
-import hmda.validation.dsl.PredicateDefaults._
+import hmda.validation.dsl.PredicateCommon._
 import hmda.validation.dsl.PredicateSyntax._
 
 object S013 extends {
