+++ validation/src/main/scala/hmda/validation/rules/ts/validity/V105.scala
@@ -3,7 +3,7 @@ package hmda.validation.rules.ts.validity
 import hmda.model.fi.ts.TransmittalSheet
 import hmda.validation.dsl.Result
 import hmda.validation.rules.EditCheck
-import hmda.validation.dsl.PredicateDefaults._
+import hmda.validation.dsl.PredicateCommon._
 import hmda.validation.dsl.PredicateSyntax._
 
 /*
