+++ validation/src/main/scala/hmda/validation/rules/ts/validity/V140.scala
@@ -4,7 +4,7 @@ import hmda.model.census.Census._
 import hmda.model.fi.ts.TransmittalSheet
 import hmda.validation.dsl.Result
 import hmda.validation.rules.EditCheck
-import hmda.validation.dsl.PredicateDefaults._
+import hmda.validation.dsl.PredicateCommon._
 import hmda.validation.dsl.PredicateSyntax._
 
 object V140 extends EditCheck[TransmittalSheet] {
